refactor: songsディレクトリに関連ファイルを入れた。呼び出し側もmoduleを介すようにした
